requirements.txt
pyserial==3.4
matplotlib==3.3.2